+++ tests/unit/test_ui_engine.py
@@ -1,6 +1,6 @@
 import yaml
 
-from ui.utils.ui_generator import compute_overrides
+from ui.utils.override_compute import compute_overrides
 from ui.utils.ui_validator import validate_inputs